Properly closing dev websocket.
