+++ java/arcs/android/devtools/DevToolsBinder.kt
@@ -15,12 +15,13 @@ import arcs.core.util.TaggedLog
 import java.lang.Exception
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.suspendCancellableCoroutine
 
 /**
  * Implementation of [IDevToolsService].
  */
 class DevToolsBinder(val scope: CoroutineScope) : IDevToolsService.Stub() {
-    private var webSocket: DevWebSocket? = null
+    private var webSocket: DevWebSocketFactory.devWebSocket? = null
     private val log = TaggedLog { "DevToolsBinder" }
 
     override fun send(str: String) {
@@ -29,19 +30,30 @@ class DevToolsBinder(val scope: CoroutineScope) : IDevToolsService.Stub() {
         }
     }
 
+    @kotlinx.coroutines.ExperimentalCoroutinesApi
     override fun start() {
         scope.launch {
             try {
-                webSocket = DevWebSocket()
-                webSocket?.start()
+                suspendCancellableCoroutine { cont ->
+                    webSocket = webSocket ?: DevWebSocketFactory().getDevWebSocket()
+                    cont.invokeOnCancellation {
+                        webSocket?.close()
+                    }
+                    webSocket?.startIfNeeded()
+                    cont.resume(Unit) {}
+                }
             } catch (e: Exception) {
-                log.debug { "Can't open Websocket. Error: [$e]." }
+                log.debug(e) { "Can't open Websocket. Error: [$e]." }
             }
         }
     }
 
-    fun destroy() {
+    override fun close() {
         webSocket?.close()
         webSocket = null
+    }
+
+    fun destroy() {
+        close()
     }
 }
